chore: separate fn for single hexagon generation
